pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "browser-policy-manager"
version = "0.1.0"
description = "Browser Policy Manager — enterprise Firefox policy editor and API service."
authors = [{ name = "Valery Ledovskoy", email = "valery@ledovskoy.com" }]
license = "MPL-2.0"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "jinja2>=3.1",
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "python-multipart>=0.0.9",
  "SQLAlchemy>=2.0",
  "aiosqlite>=0.20",
  "alembic>=1.13",
  "jsonschema>=4.23",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "httpx>=0.27",
  "ruff>=0.6",
  "black>=24.10",
  "mypy>=1.11",
  "coverage>=7.6",
]

[project.urls]
Homepage = "https://github.com/Goudron/browser-policy-manager"
Documentation = "https://github.com/Goudron/browser-policy-manager/wiki"
Source = "https://github.com/Goudron/browser-policy-manager"
Issues = "https://github.com/Goudron/browser-policy-manager/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "alembic*", "docs*", "scripts*"]

[tool.setuptools.package-data]
app = ["schemas/*.json", "schemas/**/*.json"]

# --- Tooling ---

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["build", "dist", ".venv", "venv"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_return_any = false
strict_optional = true
disallow_untyped_defs = true
show_error_context = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"
testpaths = ["tests"]
markers = [
  "order: custom ordering marker used for migration tests",
]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true
